+++ src/main/kotlin/com/yourssu/assignmentblog/domain/user/repository/UserRepository.kt
@@ -1,10 +1,40 @@
 package com.yourssu.assignmentblog.domain.user.repository
 
 import com.yourssu.assignmentblog.domain.user.domain.User
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+interface UserJpaRepository: JpaRepository<User, Long?> {
+    fun findByEmail(email: String): User?
+    fun findByRefreshToken(refreshToken: String): User?
+}
+
 
 interface UserRepository {
     fun findByEmail(email: String): User?
     fun findByRefreshToken(refreshToken: String): User?
     fun save(user: User): User
     fun delete(user: User)
+}
+@Repository
+class UserRepositoryImpl(
+    private val userJpaRepository: UserJpaRepository
+): UserRepository {
+
+    override fun findByEmail(email: String): User? {
+        return userJpaRepository.findByEmail(email)
+    }
+
+    override fun findByRefreshToken(refreshToken: String): User? {
+        return userJpaRepository.findByRefreshToken(refreshToken)
+    }
+
+    override fun save(user: User): User {
+        return userJpaRepository.save(user)
+    }
+
+    override fun delete(user: User) {
+        userJpaRepository.delete(user)
+    }
+
 }
